+++ src/shinestacker/retouch/image_viewer.py
@@ -140,7 +140,7 @@ class ImageViewer(QWidget):
             st.end_copy_brush_area_requested.connect(end_copy_brush_area)
             st.brush_size_change_requested.connect(handle_brush_size_change)
             st.brush_hardness_change_requested.connect(handle_brush_hardness_change)
-            st.brush_opcity_change_requested.connect(handle_brush_opacity_change)
+            st.brush_opacity_change_requested.connect(handle_brush_opacity_change)
             st.brush_flow_change_requested.connect(handle_brush_flow_change)
             st.needs_update_requested.connect(handle_needs_update)
             st.setFocusPolicy(Qt.StrongFocus)